requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
BeautifulSoup4==4.12.2
requests==2.31.0
schedule==1.2.1
plotly==5.18.0
python-dotenv==1.0.0
selenium==4.16.0
webdriver-manager==4.0.1
